+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.3.0.0
 
+* Add `WithdrawalsExceedAccountBalance` to `DijkstraLedgerPredFailure`
+* Removed `DijkstraSpendingOutputFromSameTx` from `DijkstraLedgerPredFailure`
+* Added `batchNonDistinctRefScriptsSize`
 * Add `guardDijkstraFeaturesForPlutusV1toV3`
 * Add `DirectDepositsNotSupported` and `AccountBalanceIntervalsNotSupported` constructors to `DijkstraContextError`
 * Add `SubTxContextError`